Module INTERNAL.Document

module DIAGS = Cobol_common.Diagnostics

module TYPES : sig ... end

include module type of struct include TYPES end

type document = TYPES.document = {

project : Cobol_lsp__.Lsp_project.TYPES.project;

textdoc : Lsp.Text_document.t;

copybook : bool;

artifacts : Cobol_parser.parsing_artifacts;

parsed : parsed_data option;

diags : DIAGS.Set.t;

}

and parsed_data = TYPES.parsed_data = {

ast : Cobol_parser.PTree.compilation_group;

cus : Cobol_data.Compilation_unit.SET.t;

definitions : Cobol_lsp__.Lsp_lookup.TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t Stdlib.Lazy.t;

references : Cobol_common.Srcloc.srcloc list Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t Stdlib.Lazy.t;

}

exception Unparseable of Lsp.Types.DocumentUri.t

Raised by retrieve_parsed_data.

exception Copybook of Lsp.Types.DocumentUri.t

type cached = TYPES.cached = {

doc_cache_filename : string;

doc_cache_checksum : Stdlib.Digest.t;

doc_cache_langid : string;

doc_cache_version : int;

doc_cache_pplog : Cobol_preproc.log;

doc_cache_tokens : Cobol_parser.tokens_with_locs;

doc_cache_comments : Cobol_preproc.comments;

doc_cache_parsed : (Cobol_parser.PTree.compilation_group * Cobol_data.Compilation_unit.SET.t) option;

doc_cache_diags : DIAGS.Set.serializable;

}

type t = document

val uri : document -> Lsp__.Uri0.t

val parse : project:Cobol_lsp__.Lsp_project.TYPES.project -> Lsp.Text_document.t -> (Cobol_parser.PTree.compilation_group option, Cobol_common.Behaviors.eidetic) Cobol_parser__.Parser_engine.parsed_result

val lazy_definitions : Cobol_parser.PTree.compilation_unit Cobol_common.with_loc list -> Cobol_data.Compilation_unit.SET.t -> Cobol_lsp__.Lsp_lookup.TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t lazy_t

val lazy_references : Cobol_parser.PTree.compilation_unit Cobol_common.with_loc list -> Cobol_data.Compilation_unit.SET.t -> Cobol_lsp__.Lsp_lookup.TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t Stdlib.Lazy.t -> Cobol_common__Srcloc.TYPES.srcloc list Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t lazy_t

val no_parsing_artifacts : Cobol_parser.parsing_artifacts

val analyze : document -> document

val blank : project:Cobol_lsp__.Lsp_project.TYPES.project -> ?copybook:bool -> Lsp.Text_document.t -> document

Creates a record for a document that is not yet parsed or analyzed.

val position_encoding : [> `UTF8 ]

val load : project:Cobol_lsp__.Lsp_project.TYPES.project -> ?copybook:bool -> Lsp__.Types.DidOpenTextDocumentParams.t -> document

val update : document -> Lsp__.Types.TextDocumentContentChangeEvent.t list -> document

val retrieve_parsed_data : document -> parsed_data

Raises Unparseable in case the document cannot be parsed entierely.

Caching utilities

val to_cache : document -> cached

val of_cache : project:Cobol_lsp__.Lsp_project.TYPES.project -> cached -> document

Raises Failure in case of bad checksum.

Miscellaneous