Module INTERNAL.Project

module TYPES : sig ... end

include module type of TYPES with type path = TYPES.path and type rootdir = TYPES.rootdir and type project = TYPES.project and type layout = TYPES.layout

type path = TYPES.path =

| RelativeToProjectRoot of string

| RelativeToFileDir of string

type rootdir = TYPES.rootdir

type project = TYPES.project = {

rootdir : rootdir;

config_checksum : Stdlib.Digest.t option;

cobol_config : Cobol_config.t;

source_format : Cobol_config.source_format_spec;

libpath : path list;

copybook_extensions : string list;

copybook_if_no_extension : bool;

}

type layout = TYPES.layout = {

project_config_filename : string;

}

type t = project

val for_ : rootdir:rootdir -> layout:layout -> t

for_ ~rootdir ~layout retrieves a project based on its root directory. This may trigger reading project configuration files if the project was not yet loaded. Any notification about the loading process is published directly via Lsp_io.send_diagnostics or Lsp_io.send_notification.

val in_existing_dir : string -> layout:layout -> t

in_existing_dir dirname ~layout retrieves a project after checking dirname actually refers to an exising directory that can serve as root for the project. The same notes as for_ apply, with the addition that Invalid_argument may is raised in case dirname is not the name of an existing directory.

val rootdir_for : uri:Lsp.Uri.t -> layout:layout -> rootdir

rootdir_for ~uri ~layout locates the project directory (that contains a file with given name layout.project_config_filename) for a file at the given URI. Returns the name of the directory that contains the file at URI if no project file is found.

val libpath_for : uri:Lsp.Uri.t -> t -> string list

libpath_for ~uri project constructs a list of directory names where copybooks are looked up, for a source file at the given URI, in the given project.

val detect_copybook : uri:Lsp.Uri.t -> t -> bool

detect_copybook ~uri project indicates whether a document at the given URI for project should be treated as a copybook.

Cached representation

type cached

val to_cache : t -> cached

val of_cache : rootdir:rootdir -> layout:layout -> cached -> t

Sets and maps

module SET : sig ... end

module MAP : Stdlib.Map.S with type key = t

Miscellaneous

val rootdir : t -> rootdir

val string_of_rootdir : rootdir -> string

val relative_path_for : uri:Lsp.Uri.t -> t -> string

val absolute_path_for : filename:string -> t -> string