Module INTERNAL.Types

module CUs = Cobol_unit.Collections.SET

module CUMap = Cobol_unit.Collections.MAP

module URIMap : sig ... end

type diagnostics = Lsp.Types.Diagnostic.t list Stdlib__Map.Make(Lsp.Uri).t

type element_at_position = {

element_at_position : element_in_context option;

enclosing_compilation_unit_name : string option;

}

Information returned by Lookup.element_at_position.

and element_in_context =

| Data_name of Cobol_ptree.qualname

| Data_full_name of Cobol_ptree.qualname

| Data_item of {

full_qn : Cobol_ptree.qualname option;

def_loc : Cobol_common.Srcloc.TYPES.srcloc;

}

| Proc_name of {

qn : Cobol_ptree.qualname;

in_section : Cobol_unit.Types.procedure_section option;

}

type name_definition = {

as_paragraph : paragraph_definition option;

as_item : item_definition option;

}

and paragraph_definition = Cobol_ptree.paragraph Cobol_common.Srcloc.TYPES.with_loc

and item_definition = Cobol_data.Types.item_definition Cobol_common.Srcloc.TYPES.with_loc

type copy_operation = string Cobol_common.Srcloc.TYPES.with_loc

with loc of COPY ...

type approx_typing_info =

| Alphanum

| Any

| Boolean

| Condition

| Group

| Index

| Numeric

| NumericEdited

| ObjectRef

| Pointer

type document = {

project : Superbol_project__.Project.TYPES.project;

textdoc : Lsp.Text_document.t;

copybook : bool;

artifacts : Cobol_parser.Outputs.artifacts;

checked : checked_doc option;

rewinder : rewinder option;

parsing_diags : Cobol_parser.Diagnostics.ALL.t;

typecking_diags : Cobol_typeck.Diagnostics.t;

}

and checked_doc = Cobol_typeck.Outputs.t

and rewinder = (Cobol_ptree.compilation_group option, Cobol_common.Behaviors.eidetic) Cobol_parser.Outputs.output Cobol_parser.rewinder

exception Unparseable of Lsp.Types.DocumentUri.t

Raised by Document.checked.

exception Copybook of Lsp.Types.DocumentUri.t

exception Internal_error of document * exn * Stdlib.Printexc.raw_backtrace

Raised by Document.load and Document.update; allows keeping consistent document contents.

type cached = {

doc_cache_filename : string;

doc_cache_checksum : Stdlib.Digest.t;

doc_cache_langid : string;

doc_cache_version : int;

doc_cache_pplog : Cobol_preproc.Trace.log;

doc_cache_tokens : Cobol_parser.Outputs.tokens_with_locs;

doc_cache_comments : Cobol_preproc.Text.comments;

doc_cache_ignored : Cobol_common.Srcloc.lexloc list;

doc_cache_checked : checked_doc option;

doc_cache_parsing_diags : Cobol_parser.Diagnostics.ALL.t;

doc_cache_typecking_diags : Cobol_typeck.Diagnostics.t;

}

Persistent representation (for caching)

include module type of Superbol_project.Config.TYPES

type path =

| RelativeToProjectRoot of string

| RelativeToFileDir of string

exception ERROR of Superbol_project__.Project_diagnostics.error

include module type of Superbol_project.TYPES

type rootdir

type project = {

rootdir : rootdir;

config : Superbol_project__.Project_config.TYPES.config;

config_filename : string;

}

type layout = {

project_config_filename : string;
(*
Name of the TOML file that is to be found at the root of each project's directory tree.
*)

relative_work_dirname : string option;
(*
Relative name of a directory where the LSP should put its working files (caches, etc). No such storage is allowed when None.
*)

rootdir_fallback_policy : rootdir_fallback_policy;
(*
Policy to determine the root directory of projects based on individual filenames.
*)

}

and rootdir_fallback_policy =

| Same_as_file_directory

| Given_directory of string

type storage =

| No_storage

| Store_in_file of {

relative_filename : string;

}

| Store_in_shared_dir of {

dirname : string;

}

type config = {

project_layout : Superbol_project__.Project.TYPES.layout;

cache_config : Cobol_lsp__.Lsp_project_cache.TYPES.config;

enable_client_configs : bool;

force_syntax_diagnostics : bool;

}

type params = {

config : config;

root_uri : Lsp.Types.DocumentUri.t option;

workspace_folders : Lsp.Types.DocumentUri.t list;

with_semantic_tokens : bool;

with_client_config_watcher : bool;

with_client_file_watcher : bool;

}

type registry = private {

projects : Cobol_lsp__.Lsp_project.SET.t;

docs : Cobol_lsp__.Lsp_document.TYPES.document Stdlib__Map.Make(Lsp.Uri).t;

indirect_diags : Lsp.Types.Diagnostic.t list Stdlib__Map.Make(Lsp.Uri).t Stdlib__Map.Make(Lsp.Uri).t;

pending_tasks : pending_tasks;

sub_state : sub_state;

params : params;

}

and sub_state

and pending_tasks

type state =

| NotInitialized of config

| Initialized of params

| Running of registry

| ShuttingDown

| Exit of exit_status

and exit_status = (unit, string) Stdlib.result

type 'a error =

| InvalidStatus of state

| UnhandledRequest of 'a Lsp.Client_request.t

| UnknownRequest of string

exception Document_not_found of Lsp.Types.TextDocumentIdentifier.t