Module INTERNAL.Document

module DIAGS = Cobol_common.Diagnostics

module TYPES : sig ... end

include module type of struct include TYPES end

type document = TYPES.document = {

project : Superbol_project__.Project.TYPES.project;

textdoc : Lsp.Text_document.t;

copybook : bool;

artifacts : Cobol_parser.Outputs.artifacts;

parsed : parsed_data option;

rewinder : rewinder option;

diags : DIAGS.Set.t;

}

and parsed_data = TYPES.parsed_data = {

ptree : Cobol_ptree.compilation_group;

cus : Cobol_data.Compilation_unit.SET.t;

definitions : Cobol_lsp__.Lsp_lookup.TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t Stdlib.Lazy.t;

references : Cobol_common.Srcloc.srcloc list Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t Stdlib.Lazy.t;

}

and rewinder = (Cobol_ptree.compilation_group option, Cobol_common.Behaviors.eidetic) Cobol_parser.Outputs.output Cobol_parser.rewinder

exception Unparseable of Lsp.Types.DocumentUri.t

Raised by Document.retrieve_parsed_data.

exception Copybook of Lsp.Types.DocumentUri.t

exception Internal_error of document * exn * Stdlib.Printexc.raw_backtrace

Raised by Document.load and Document.update; allows keeping consistent document contents.

type cached = TYPES.cached = {

doc_cache_filename : string;

doc_cache_checksum : Stdlib.Digest.t;

doc_cache_langid : string;

doc_cache_version : int;

doc_cache_pplog : Cobol_preproc.Trace.log;

doc_cache_tokens : Cobol_parser.Outputs.tokens_with_locs;

doc_cache_comments : Cobol_preproc.Text.comments;

doc_cache_ignored : Cobol_common.Srcloc.lexloc list;

doc_cache_parsed : (Cobol_ptree.compilation_group * Cobol_data.Compilation_unit.SET.t) option;

doc_cache_diags : DIAGS.Set.serializable;

}

type t = document

val uri : document -> Lsp__.Uri0.t

val rewindable_parse : document -> ((Cobol_ptree.compilation_group option, Cobol_common.Behaviors.eidetic) Cobol_parser__.Parser_outputs.output * (Cobol_ptree.compilation_group option, Cobol_common.Behaviors.eidetic) Cobol_parser__.Parser_outputs.output Cobol_parser.rewinder) Cobol_common.Diagnostics.with_diags

val lazy_definitions : Cobol_ptree.compilation_group -> Cobol_data.Compilation_unit.SET.t -> Cobol_lsp__.Lsp_lookup.TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t lazy_t

val lazy_references : Cobol_ptree.compilation_group -> Cobol_data.Compilation_unit.SET.t -> Cobol_lsp__.Lsp_lookup.TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t Stdlib.Lazy.t -> Cobol_common__Srcloc.TYPES.srcloc list Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t lazy_t

val no_artifacts : Cobol_parser.Outputs.artifacts

val gather_parsed_data : 'a Cobol_parser.Outputs.parsed_compilation_group -> parsed_data option DIAGS.with_diags

val extract_parsed_infos : document -> ((Cobol_ptree.compilation_group option, Cobol_common.Behaviors.eidetic) Cobol_parser__.Parser_outputs.output * rewinder) DIAGS.with_diags -> document

val parse_and_analyze : document -> document

val reparse_and_analyze : ?position:Cobol_parser.position -> document -> document

val blank : project:Superbol_project__.Project.TYPES.project -> ?copybook:bool -> Lsp.Text_document.t -> document

Creates a record for a document that is not yet parsed or analyzed.

val position_encoding : [> `UTF8 ]

val load : project:Superbol_project__.Project.TYPES.project -> ?copybook:bool -> Lsp__.Types.DidOpenTextDocumentParams.t -> document

val first_change_pos : Lsp.Types.TextDocumentContentChangeEvent.t list -> Cobol_parser.position

val update : document -> Lsp__.Types.TextDocumentContentChangeEvent.t list -> document

val retrieve_parsed_data : document -> parsed_data

Raises Unparseable in case the document cannot be parsed entierely, or Copybook in case the document is not a main program.

Caching utilities

val to_cache : document -> cached

val of_cache : project:Superbol_project__.Project.TYPES.project -> cached -> document

Raises Failure in case of bad checksum.

Miscellaneous