Module INTERNAL.Document

module DIAGS = Cobol_common.Diagnostics

module TYPES : sig ... end

include module type of struct include TYPES end

type document = TYPES.document = {

project : Superbol_project__.Project.TYPES.project;

textdoc : Lsp.Text_document.t;

copybook : bool;

artifacts : Cobol_parser.Outputs.artifacts;

checked : checked_doc option;

rewinder : rewinder option;

parsing_diags : Cobol_parser.Diagnostics.ALL.t;

typecking_diags : Cobol_typeck.Diagnostics.t;

}

and checked_doc = Cobol_typeck.Outputs.t

and rewinder = (Cobol_ptree.compilation_group option, Cobol_common.Behaviors.eidetic) Cobol_parser.Outputs.output Cobol_parser.rewinder

exception Unparseable of Lsp.Types.DocumentUri.t

Raised by Document.checked.

exception Copybook of Lsp.Types.DocumentUri.t

exception Internal_error of document * exn * Stdlib.Printexc.raw_backtrace

Raised by Document.load and Document.update; allows keeping consistent document contents.

type cached = TYPES.cached = {

doc_cache_filename : string;

doc_cache_checksum : Stdlib.Digest.t;

doc_cache_langid : string;

doc_cache_version : int;

doc_cache_pplog : Cobol_preproc.Trace.log;

doc_cache_tokens : Cobol_parser.Outputs.tokens_with_locs;

doc_cache_comments : Cobol_preproc.Text.comments;

doc_cache_ignored : Cobol_common.Srcloc.lexloc list;

doc_cache_checked : checked_doc option;

doc_cache_parsing_diags : Cobol_parser.Diagnostics.ALL.t;

doc_cache_typecking_diags : Cobol_typeck.Diagnostics.t;

}

Persistent representation (for caching)

type t = document

val uri : document -> Lsp__.Uri0.t

val rewindable_parse : document -> ((Cobol_ptree.compilation_group option, Cobol_common.Behaviors.eidetic) Cobol_parser__.Parser_outputs.output * (Cobol_ptree.compilation_group option, Cobol_common.Behaviors.eidetic) Cobol_parser__.Parser_outputs.output Cobol_parser.rewinder) Cobol_common__Diagnostics_accumulator.MAKE(Cobol_parser__.Parser_diagnostics.ALL).TYPES.with_diags

val no_artifacts : Cobol_parser.Outputs.artifacts

val check : document -> (Cobol_common.Behaviors.eidetic Cobol_parser.Outputs.parsed_compilation_group * rewinder) Cobol_parser.Outputs.with_diags -> document

val parse_and_analyze : document -> document

val reparse_and_analyze : ?position:Cobol_parser.position -> document -> document

val blank : project:Superbol_project__.Project.TYPES.project -> ?copybook:bool -> Lsp.Text_document.t -> document

Creates a record for a document that is not yet parsed or analyzed.

val position_encoding : [> `UTF8 ]

val load : project:Superbol_project__.Project.TYPES.project -> ?copybook:bool -> Lsp__.Types.DidOpenTextDocumentParams.t -> document

val first_change_pos : Lsp.Types.TextDocumentContentChangeEvent.t list -> Cobol_parser.position

val reload : document -> document

val update : document -> Lsp__.Types.TextDocumentContentChangeEvent.t list -> document

val checked : document -> checked_doc

Raises Unparseable in case the document cannot be parsed entierely, or Copybook in case the document is not a main program.

val diagnostics : document -> DIAGS.Set.t

Caching utilities

val to_cache : document -> cached

val of_cache : project:Superbol_project__.Project.TYPES.project -> cached -> document

Raises Failure in case of bad checksum.

Miscellaneous