Module INTERNAL.Lookup

module TYPES : sig ... end

Type definitions to help representing and querying (partially) parsed COBOL programs.

val bare : Cobol_ptree.name Cobol_common.Srcloc.with_loc -> Cobol_ptree.qualname

val qual : Cobol_ptree.name Cobol_common.Srcloc.with_loc -> Cobol_ptree.qualname option -> Cobol_ptree.qualname

val simple_name : Cobol_ptree.qualname -> string

val baseloc_of_qualname : Cobol_ptree.qualname -> Cobol_common.srcloc

val lexloc_of_qualname_in : filename:string -> Cobol_ptree.qualname -> Cobol_common.Srcloc.lexloc

lexloc_of_qualname_in ~filename qualname returns the full location of qualname, from the first qualifier to the last. This function is temporary and is expected to be replaced once a better way of getting this location is found.

val qualname_at_pos : filename:string -> Cobol_ptree.qualname -> Lsp.Types.Position.t -> Cobol_ptree.qualname

qualname_at_pos ~filename qualname pos returns the qualname built from all the qualifiers of qualname that are after or at position pos ion filename. This function is temporary and is expected to be replaced once a better way of finding the qualname is implemented.

val name_of_compunit : Cobol_ptree.compilation_unit Cobol_common.with_loc -> string

val compilation_unit_by_name : Cobol_ptree.name -> Cobol_ptree.compilation_group -> Cobol_ptree.compilation_unit Cobol_common.with_loc option

val names_at_position : uri:Lsp.Uri.t -> Lsp.Types.Position.t -> Cobol_ptree__.Compilation_group.compilation_group -> TYPES.names_at_position

names_at_position pos ptree seeks the compilation unit name and qualified name at the given position pos, in compilation group parse-tree ptree.

Name definitions retrieval

val def_binding : Cobol_ptree.qualname -> TYPES.name_definition Cobol_data.Qualmap.t -> TYPES.name_definition

val def_paragraph : Cobol_ptree.qualname -> TYPES.paragraph_definition -> TYPES.name_definition Cobol_data.Qualmap.t -> TYPES.name_definition Cobol_data.Qualmap.t

val def_item : Cobol_ptree.qualname -> TYPES.item_definition -> TYPES.name_definition Cobol_data.Qualmap.t -> TYPES.name_definition Cobol_data.Qualmap.t

val def_redefinition : Cobol_ptree.qualname -> Cobol_ptree.name Cobol_common.Srcloc.with_loc -> TYPES.name_definition Cobol_data.Qualmap.t -> TYPES.name_definition Cobol_data.Qualmap.t

val definitions : Cobol_data.Types.compilation_unit -> TYPES.name_definitions_in_compilation_unit

Creates a mapping of item definitions from a type-checked group item.

val update_definitions_based_on_compilation_group_ptree : f: (Cobol_ptree.compilation_unit Cobol_common.with_loc -> 'a Cobol_data.Qualmap.t -> 'a Cobol_data.Qualmap.t) -> Cobol_ptree.compilation_group -> 'a Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t -> 'a Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t

val add_rename_item_definitions : Cobol_ptree.compilation_group -> TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t -> TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t

val add_paragraph_definitions : Cobol_ptree.compilation_group -> TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t -> TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t

val add_redefine_definitions : Cobol_ptree.compilation_group -> TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t -> TYPES.name_definition Cobol_data.Qualmap.t Cobol_data.Compilation_unit.MAP.t

val references_of_qualname : Cobol_ptree.qualname -> Cobol_ptree__.Compilation_group.compilation_unit Cobol_common.Srcloc.TYPES.with_loc -> TYPES.name_definition Cobol_data.Qualmap.t -> Cobol_common__Srcloc.TYPES.srcloc list

val references : TYPES.name_definition Cobol_data.Qualmap.t -> Cobol_ptree__.Compilation_group.compilation_unit Cobol_common.Srcloc.TYPES.with_loc -> Cobol_common__Srcloc.TYPES.srcloc list Cobol_data.Qualmap.t

val copy_at_pos : filename:string -> Lsp.Types.Position.t -> Cobol_ptree__.Compilation_group.compilation_group -> TYPES.copy_operation option